* src/gpgme.c: Fix implicit conversion

+++ src/gpgme.c
@@ -282,7 +282,7 @@ int wget_verify_job(JOB *job, wget_http_response *resp, wget_gpg_info_t *info)
 		return WGET_E_INVALID;
 	}
 
-	size_t num_bytes = -1;
+	size_t num_bytes = (size_t) -1;
 	char *file_contents = NULL;
 	debug_printf("Verifying %s against sig %s\n", corrected_base_file, job->sig_filename);
